requirements.txt
asgiref==3.11.0
boto3==1.42.35
botocore==1.42.35
Django==6.0.1
django-environ==0.12.0
django-storages==1.14.6
jmespath==1.1.0
pillow==12.1.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
django-nested-admin==4.0.2
djangorestframework==3.14.0
urllib3==2.6.3
